+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed a `case` bug.
 - Added some testing scripts.
 - Carets are now allowed in identifiers.
+- Fixed `to block!` for strings.
 
 
 # Version 0.0.9